Finished the conversion to Allegro 5! It doesn't really WORK yet, but it compiles... There's also many "NEW_ALLEGRO" markers to note where I just commented something out instead of converting it to Allegro 5 (such as the sound system).

+++ src/Sprite.cpp
@@ -33,14 +33,14 @@ void Sprite::update() {
       fudge += speed;
       
       while (fudge >= GAME_TICKER) {
-        if (visualX != x * getTileSize()) {
+        if ((int)visualX != x * getTileSize()) {
           if (visualX < x * getTileSize()) {
             visualX++;
           } else {
             visualX--;
           }
         }
-        if (visualY != y * getTileSize()) {
+        if ((int)visualY != y * getTileSize()) {
           if (visualY < y * getTileSize()) {
             visualY++;
           } else {
